+++ contracts/crydr/jnt/JNTControllerInterface.sol
@@ -11,10 +11,10 @@ contract JNTControllerInterface {
 
   /* Events */
 
-  event JNTChargedEvent(address indexed payableservice, address indexed from, address indexed to, uint value);
+  event JNTChargedEvent(address indexed payableservice, address indexed from, address indexed to, uint256 value);
 
 
   /* Actions */
 
-  function chargeJNT(address _from, address _to, uint _value);
+  function chargeJNT(address _from, address _to, uint256 _value);
 }
